FEAT: implement testing
